requirements.txt
chess==1.7.0
Django==4.0
djangorestframework==3.13.0
gunicorn==20.1.0
